Move agent tool implementations to submodules
