+++ server/bleep/src/webserver/answer/conversations.rs
@@ -9,10 +9,11 @@ use std::{fmt, str::FromStr};
 use tracing::info;
 
 use crate::{
+    agent::exchange::Exchange,
     db::SqlDb,
     repo::RepoRef,
     webserver::{self, middleware::User, Error, ErrorKind},
-    Application, agent::exchange::Exchange,
+    Application,
 };
 
 type Conversation = (RepoRef, Vec<Exchange>);
